Rename and simplify frontend component names
