+++ client/task-list/tasks/payments/PaymentGatewaySuggestions/components/List/index.js
@@ -1,0 +1,2 @@
+export { List } from './List';
+export { Placeholder } from './Placeholder';
