+++ client/task-list/tasks/payments/PaymentGatewaySuggestions/components/Setup/index.js
@@ -1,0 +1,2 @@
+export { Setup } from './Setup';
+export { Placeholder } from './Placeholder';
