--- client/task-list/tasks/payments/RemotePayments/components/PaymentMethod/index.js
@@ -1,2 +1,0 @@
-export { PaymentMethod } from './PaymentMethod';
-export { PaymentMethodPlaceholder } from './PaymentMethodPlaceholder';
