--- client/task-list/tasks/payments/RemotePayments/components/RecommendedPaymentGatewayList/index.js
@@ -1,2 +1,0 @@
-export { RecommendedPaymentGatewayList } from './RecommendedPaymentGatewayList';
-export { RecommendedPaymentGatewayListPlaceholder } from './RecommendedPaymentGatewayListPlaceholder';
