+++ client/task-list/tasks/payments/index.js
@@ -2,11 +2,11 @@
  * Internal dependencies
  */
 import { LocalPayments } from './LocalPayments';
-import { RemotePayments } from './RemotePayments';
+import { PaymentGatewaySuggestions } from './PaymentGatewaySuggestions';
 
 export const Payments = ( { query } ) => {
 	if ( window.wcAdminFeatures[ 'remote-payment-methods' ] ) {
-		return <RemotePayments query={ query } />;
+		return <PaymentGatewaySuggestions query={ query } />;
 	}
 
 	return <LocalPayments query={ query } />;
